Changed the Building model to be in line with the client.
Removed the x and y fields. One tile buildings will now use the firstX and firstY fields for their only field.
Added guardCount and buildPoints fields.
Adjusted string represenation to take into account that harbors and bridges now are defined by the two fields they are on, just like streets.
Added the BP display to the string representations of production buildings and walls.
Added the guardCount display to the string representation of walls.